remove tab-related code for tabbed notebook, add tooltips to new buttons

+++ gtk2_ardour/ardour_ui_dialogs.cc
@@ -367,7 +367,7 @@ ARDOUR_UI::detach_tabbable (Tabbable* t)
 void
 ARDOUR_UI::tabs_switch (GtkNotebookPage*, guint page)
 {
-	if (page == _tabs.page_num (editor->contents())) {
+	if (page == (guint) _tabs.page_num (editor->contents())) {
 		editor_visibility_button.set_active_state (Gtkmm2ext::ImplicitActive);
 		if (mixer && (mixer->tabbed() || mixer->tabbed_by_default())) {
 			mixer_visibility_button.set_active_state (Gtkmm2ext::Off);
@@ -375,7 +375,7 @@ ARDOUR_UI::tabs_switch (GtkNotebookPage*, guint page)
 		if (rc_option_editor && (rc_option_editor->tabbed() || rc_option_editor->tabbed_by_default())) {
 			prefs_visibility_button.set_active_state (Gtkmm2ext::Off);
 		}
-	} else if (page == _tabs.page_num (mixer->contents())) {
+	} else if (page == (guint) _tabs.page_num (mixer->contents())) {
 		if (editor && (editor->tabbed() || editor->tabbed_by_default())) {
 			editor_visibility_button.set_active_state (Gtkmm2ext::Off);
 		}
